requirements.txt
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.9
django-ckeditor==6.2.0
django-heroku==0.3.1
django-import-export==2.6.1
django-js-asset==1.2.2
django-simple-history==3.0.0
et-xmlfile==1.1.0
gunicorn==20.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.4.0
psycopg2==2.9.2
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
tablib==3.1.0
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
